Adding RPC Backend
Since flask doesn't support asynchronous processing, we need to run a server
that does.

The model of flask is to use a single thread to perform all the operations for
a given request. This allows them to spawn as many threads as they want for
every request. It scales well, but in this case, we don't actually want it to
work that way.

I've decided to simply have an external daemon process running that performs
the scheduling.

The flask front-end passes the information to the processing back-end, which
performs the processing and returns the result. Basically like a function call
that may take a long time. Then the flask server can actually work as our
autonomic manager as well, tuning the CPU scheduling back-end server through
the database. The flask server can determine when more plans are required, can
determine when it needs to limit the number of request, etc by looking in the
database and asking the scheduler how many jobs it currently has.