No Hang Server
The server would hang on infinite loops or on programs that didn't output
anything. It no longer does this and is able to correctly terminate such
programs.

Furthermore, throwing errors used to lead to the program hanging, but is fixed.
The error message for erroneous programs is not helpful though.